added prepare_base_model files

+++ chicken_disease_classification/src/ChickenDiseaseClassification/pipeline/stage_01_data_ingestion.py
@@ -1,3 +1,4 @@
+from xml.etree.ElementInclude import LimitedRecursiveIncludeError
 from ChickenDiseaseClassification.config.configuration import ConfigurationManager
 from ChickenDiseaseClassification.components.data_ingestion import DataIngestion
 from ChickenDiseaseClassification import logger
@@ -8,7 +9,10 @@ class DataIngestionTrainingPipeline:
     
     def main(self):
         config = ConfigurationManager()
-        data_ingestion_config = config.get_data_ingestion_config()
-        data_ingestion = DataIngestion(data_ingestion_config = data_ingestion_config)
-        data_ingestion.copy_file()
-        data_ingestion.extract_zip_file()
+        if not config.is_exists:
+            data_ingestion_config = config.get_data_ingestion_config()
+            data_ingestion = DataIngestion(data_ingestion_config = data_ingestion_config)
+            data_ingestion.copy_file()
+            data_ingestion.extract_zip_file()
+        else:
+            logger.info("The folders and necessary files are  already present")
